Fix job.wait() issue on Windows
It seems 'job_status()' cannot return a correct status in loop on Windows.
To fix that, call 'sleep 1m' on every 10ms to let OS to interrupt the Vim's
thread.

Additionally, it seems the previous implementation sometime failed to retrieve
entire data when the process has terminated soon with huge data. This new
implementation may fixed that issue as well.